feature(components): moving all renderables to components
