+++ package.json
@@ -1,40 +1,43 @@
 {
-  "name": "fast-food",
-  "version": "1.0.0",
-  "description": "**FastFood** is an explorative experiment into web tecnologies for interactivity, the game is based in *javascript* and do not include any library or framework. The game also implies a code hierarchy and organization keeping the abstraction for more escalability, OOP is being implemented in the code, for more efficiency.",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "tsc": "tsc --noEmit --watch",
-    "lint": "eslint . --no-warn-ignored",
-    "dev": "vite",
-    "build": "vite build",
-    "prepare": "husky install"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Pheetbag/fast-food.git"
-  },
-  "author": "Pedro Jimenez <pjjm.jimenez@gmail.com>",
-  "license": "LGPL-3.0-or-later",
-  "bugs": {
-    "url": "https://github.com/Pheetbag/fast-food/issues"
-  },
-  "homepage": "https://github.com/Pheetbag/fast-food#readme",
-  "packageManager": "pnpm@10.15.0+sha512.486ebc259d3e999a4e8691ce03b5cac4a71cbeca39372a9b762cb500cfdf0873e2cb16abe3d951b1ee2cf012503f027b98b6584e4df22524e0c7450d9ec7aa7b",
-  "devDependencies": {
-    "@eslint/js": "^9.34.0",
-    "eslint": "^9.34.0",
-    "globals": "^16.3.0",
-    "husky": "^9.1.7",
-    "lint-staged": "^16.1.5",
-    "prettier": "3.6.2",
-    "typescript": "^5.9.2",
-    "typescript-eslint": "^8.40.0",
-    "vite": "^7.1.3"
-  },
-  "lint-staged": {
-    "*.css": "prettier --write",
-    "*.{js,ts,mjs}": ["prettier --write", "eslint --fix --no-warn-ignored"]
-  }
+    "name": "fast-food",
+    "version": "1.0.0",
+    "description": "**FastFood** is an explorative experiment into web tecnologies for interactivity, the game is based in *javascript* and do not include any library or framework. The game also implies a code hierarchy and organization keeping the abstraction for more escalability, OOP is being implemented in the code, for more efficiency.",
+    "main": "index.js",
+    "scripts": {
+        "test": "echo \"Error: no test specified\" && exit 1",
+        "tsc": "tsc --noEmit --watch",
+        "lint": "eslint . --no-warn-ignored",
+        "dev": "vite",
+        "build": "vite build",
+        "prepare": "husky install"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Pheetbag/fast-food.git"
+    },
+    "author": "Pedro Jimenez <pjjm.jimenez@gmail.com>",
+    "license": "LGPL-3.0-or-later",
+    "bugs": {
+        "url": "https://github.com/Pheetbag/fast-food/issues"
+    },
+    "homepage": "https://github.com/Pheetbag/fast-food#readme",
+    "packageManager": "pnpm@10.26.2+sha512.0e308ff2005fc7410366f154f625f6631ab2b16b1d2e70238444dd6ae9d630a8482d92a451144debc492416896ed16f7b114a86ec68b8404b2443869e68ffda6",
+    "devDependencies": {
+        "@eslint/js": "^9.34.0",
+        "eslint": "^9.34.0",
+        "globals": "^16.3.0",
+        "husky": "^9.1.7",
+        "lint-staged": "^16.1.5",
+        "prettier": "3.6.2",
+        "typescript": "^5.9.2",
+        "typescript-eslint": "^8.40.0",
+        "vite": "^7.1.3"
+    },
+    "lint-staged": {
+        "*.css": "prettier --write",
+        "*.{js,ts,mjs}": [
+            "prettier --write",
+            "eslint --fix --no-warn-ignored"
+        ]
+    }
 }
