+++ pnpm-workspace.yaml
@@ -1,2 +1,2 @@
 onlyBuiltDependencies:
-  - esbuild
+    - esbuild
