+++ styles/defaults.css
@@ -1,215 +1,215 @@
 *,
 *:after,
 *:before {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: "Roboto", sans-serif;
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+    font-family: "Roboto", sans-serif;
 }
 
 body,
 html {
-  background-color: black;
-  color: white;
-  height: 100%;
+    background-color: black;
+    color: white;
+    height: 100%;
 }
 
 .window {
-  height: 100%;
-  width: 100%;
-  position: absolute;
+    height: 100%;
+    width: 100%;
+    position: absolute;
 }
 
 #preloaderEngine {
-  visibility: hidden;
-  position: absolute;
-  top: 0px;
-  width: 0px;
-  height: 0px;
+    visibility: hidden;
+    position: absolute;
+    top: 0px;
+    width: 0px;
+    height: 0px;
 }
 
 .gameTitle,
 h1 {
-  font-weight: 400;
-  font-family: "Chango", cursive;
-  font-size: 30px;
-  text-shadow:
-    -2px -2px 1px #111111,
-    2px 2px 1px #111,
-    -2px 2px 1px #111,
-    2px -2px 1px #111;
+    font-weight: 400;
+    font-family: "Chango", cursive;
+    font-size: 30px;
+    text-shadow:
+        -2px -2px 1px #111111,
+        2px 2px 1px #111,
+        -2px 2px 1px #111,
+        2px -2px 1px #111;
 }
 
 .btn-general {
-  cursor: pointer;
+    cursor: pointer;
 
-  display: inline-block;
+    display: inline-block;
 
-  background-color: #1e526e;
+    background-color: #1e526e;
 
-  border-radius: 5px;
-  border: 2px solid #0c3645;
+    border-radius: 5px;
+    border: 2px solid #0c3645;
 }
 
 .btn-general:hover {
-  background-color: #12425c;
+    background-color: #12425c;
 }
 
 .btn-general:active {
-  background-color: #032e46;
+    background-color: #032e46;
 }
 
 .btn-selecter {
-  width: 35px;
-  height: 35px;
+    width: 35px;
+    height: 35px;
 }
 
 .btn-writeon {
-  font-family: "Roboto", sans-serif;
-  font-weight: 700;
-  padding: 3px 15px;
-  height: 35px;
+    font-family: "Roboto", sans-serif;
+    font-weight: 700;
+    padding: 3px 15px;
+    height: 35px;
 }
 
 .btn-close {
-  background-image: url(/assets/ui_ico/closeico.png);
-  background-size: 15px;
-  background-position: center;
-  background-repeat: no-repeat;
+    background-image: url(/assets/ui_ico/closeico.png);
+    background-size: 15px;
+    background-position: center;
+    background-repeat: no-repeat;
 }
 
 .btn-pause {
-  background-image: url(/assets/ui_ico/pause.png);
-  background-position: center;
-  background-repeat: no-repeat;
-  background-size: 15px;
+    background-image: url(/assets/ui_ico/pause.png);
+    background-position: center;
+    background-repeat: no-repeat;
+    background-size: 15px;
 }
 
 .btn-gray,
 .btn-disable {
-  color: white;
+    color: white;
 
-  background: #949494;
-  background: linear-gradient(
-    to bottom,
-    #949494 0%,
-    #949494 3%,
-    #707070 5%,
-    #707070 44%,
-    #636363 47%,
-    #6b6b6b 100%
-  );
+    background: #949494;
+    background: linear-gradient(
+        to bottom,
+        #949494 0%,
+        #949494 3%,
+        #707070 5%,
+        #707070 44%,
+        #636363 47%,
+        #6b6b6b 100%
+    );
 
-  border: 1px solid black;
-  border-radius: 7px;
+    border: 1px solid black;
+    border-radius: 7px;
 
-  cursor: pointer;
+    cursor: pointer;
 }
 
 .btn-green,
 .btn-good {
-  color: white;
-  text-shadow:
-    -2px -2px 1px #2c6200,
-    2px 2px 1px #2c6200,
-    -2px 2px 1px #2c6200,
-    2px -2px 1px #2c6200;
+    color: white;
+    text-shadow:
+        -2px -2px 1px #2c6200,
+        2px 2px 1px #2c6200,
+        -2px 2px 1px #2c6200,
+        2px -2px 1px #2c6200;
 
-  background: #d4fd81;
-  background: linear-gradient(
-    to bottom,
-    #d4fd81 0%,
-    #d4fd81 4%,
-    #94d801 7%,
-    #94d801 46%,
-    #87cb00 53%,
-    #87cb00 92%,
-    #619100 94%,
-    #619100 100%
-  );
+    background: #d4fd81;
+    background: linear-gradient(
+        to bottom,
+        #d4fd81 0%,
+        #d4fd81 4%,
+        #94d801 7%,
+        #94d801 46%,
+        #87cb00 53%,
+        #87cb00 92%,
+        #619100 94%,
+        #619100 100%
+    );
 
-  border: 1px solid #243806;
-  border-radius: 7px;
+    border: 1px solid #243806;
+    border-radius: 7px;
 
-  cursor: pointer;
+    cursor: pointer;
 }
 
 .btn-green:hover,
 .btn-good:hover {
-  background: #bdfc40;
-  background: linear-gradient(
-    to bottom,
-    #bdfc40 0%,
-    #bdfc40 4%,
-    #7ab201 7%,
-    #7ab201 46%,
-    #70a800 53%,
-    #70a800 92%,
-    #4e7500 94%,
-    #4e7500 100%
-  );
+    background: #bdfc40;
+    background: linear-gradient(
+        to bottom,
+        #bdfc40 0%,
+        #bdfc40 4%,
+        #7ab201 7%,
+        #7ab201 46%,
+        #70a800 53%,
+        #70a800 92%,
+        #4e7500 94%,
+        #4e7500 100%
+    );
 }
 
 .btn-green:active,
 .btn-good:active {
-  background: rgba(75, 112, 0, 1);
-  background: linear-gradient(
-    to bottom,
-    rgba(75, 112, 0, 1) 0%,
-    rgba(75, 112, 0, 1) 3%,
-    rgba(109, 163, 0, 1) 5%,
-    rgba(109, 163, 0, 1) 44%,
-    rgba(118, 173, 1, 1) 47%,
-    rgba(118, 173, 1, 1) 100%
-  );
+    background: rgba(75, 112, 0, 1);
+    background: linear-gradient(
+        to bottom,
+        rgba(75, 112, 0, 1) 0%,
+        rgba(75, 112, 0, 1) 3%,
+        rgba(109, 163, 0, 1) 5%,
+        rgba(109, 163, 0, 1) 44%,
+        rgba(118, 173, 1, 1) 47%,
+        rgba(118, 173, 1, 1) 100%
+    );
 }
 
 .btn-orange {
-  color: white;
+    color: white;
 
-  border: 1px solid #243806;
-  border-radius: 7px;
+    border: 1px solid #243806;
+    border-radius: 7px;
 
-  cursor: pointer;
+    cursor: pointer;
 
-  background: rgba(235, 113, 38, 1);
-  background: linear-gradient(
-    to bottom,
-    rgba(235, 113, 38, 1) 0%,
-    rgba(235, 113, 38, 1) 3%,
-    rgba(199, 82, 4, 1) 4%,
-    rgba(199, 82, 4, 1) 50%,
-    rgba(184, 68, 6, 1) 51%,
-    rgba(184, 68, 6, 1) 94%,
-    rgba(125, 46, 6, 1) 95%,
-    rgba(125, 46, 6, 1) 100%
-  );
+    background: rgba(235, 113, 38, 1);
+    background: linear-gradient(
+        to bottom,
+        rgba(235, 113, 38, 1) 0%,
+        rgba(235, 113, 38, 1) 3%,
+        rgba(199, 82, 4, 1) 4%,
+        rgba(199, 82, 4, 1) 50%,
+        rgba(184, 68, 6, 1) 51%,
+        rgba(184, 68, 6, 1) 94%,
+        rgba(125, 46, 6, 1) 95%,
+        rgba(125, 46, 6, 1) 100%
+    );
 }
 
 .btn-orange:hover {
-  background: rgba(235, 113, 38, 1);
-  background: linear-gradient(
-    to bottom,
-    rgba(235, 113, 38, 1) 0%,
-    rgba(235, 113, 38, 1) 3%,
-    rgba(148, 64, 8, 1) 4%,
-    rgba(148, 64, 8, 1) 50%,
-    rgba(122, 45, 6, 1) 51%,
-    rgba(122, 45, 6, 1) 94%,
-    rgba(87, 32, 5, 1) 95%,
-    rgba(87, 32, 5, 1) 100%
-  );
+    background: rgba(235, 113, 38, 1);
+    background: linear-gradient(
+        to bottom,
+        rgba(235, 113, 38, 1) 0%,
+        rgba(235, 113, 38, 1) 3%,
+        rgba(148, 64, 8, 1) 4%,
+        rgba(148, 64, 8, 1) 50%,
+        rgba(122, 45, 6, 1) 51%,
+        rgba(122, 45, 6, 1) 94%,
+        rgba(87, 32, 5, 1) 95%,
+        rgba(87, 32, 5, 1) 100%
+    );
 }
 
 .btn-orange:active {
-  background: rgba(87, 32, 5, 1);
-  background: linear-gradient(
-    to bottom,
-    rgba(87, 32, 5, 1) 0%,
-    rgba(87, 32, 5, 1) 6%,
-    rgba(148, 64, 8, 1) 8%,
-    rgba(122, 45, 6, 1) 50%,
-    rgba(148, 64, 8, 1) 52%,
-    rgba(148, 64, 8, 1) 100%
-  );
+    background: rgba(87, 32, 5, 1);
+    background: linear-gradient(
+        to bottom,
+        rgba(87, 32, 5, 1) 0%,
+        rgba(87, 32, 5, 1) 6%,
+        rgba(148, 64, 8, 1) 8%,
+        rgba(122, 45, 6, 1) 50%,
+        rgba(148, 64, 8, 1) 52%,
+        rgba(148, 64, 8, 1) 100%
+    );
 }
